add wait group wait timeout function to util
