+++ go/common/async/README.md
@@ -1,0 +1,1 @@
+Utilities for asynchronous programming (go routines, wait groups, etc.)
