+++ integration/ethereummock/mock_l1_network.go
@@ -2,6 +2,7 @@ package ethereummock
 
 import (
 	"fmt"
+	"github.com/obscuronet/go-obscuro/go/common/async"
 	"time"
 
 	"github.com/obscuronet/go-obscuro/go/common/log"
@@ -48,7 +49,7 @@ func (n *MockEthNetwork) BroadcastBlock(b common.EncodedL1Block, p common.Encode
 	for _, m := range n.AllNodes {
 		if m.Info().L2ID != n.CurrentNode.Info().L2ID {
 			t := m
-			common.Schedule(n.delay(), func() { t.P2PReceiveBlock(b, p) })
+			async.Schedule(n.delay(), func() { t.P2PReceiveBlock(b, p) })
 		} else {
 			m.logger.Info(printBlock(bl, m))
 		}
@@ -65,7 +66,7 @@ func (n *MockEthNetwork) BroadcastTx(tx *types.Transaction) {
 			// the time to broadcast a tx is half that of a L1 block, because it is smaller.
 			// todo - find a better way to express this
 			d := n.delay() / 2
-			common.Schedule(d, func() { t.P2PGossipTx(tx) })
+			async.Schedule(d, func() { t.P2PGossipTx(tx) })
 		}
 	}
 }
